Renamed management resource containers
