+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/management/resources/ManagementAttachmentResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.management.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.management.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractAttachmentResourceContainer;
 
 @ApplicationScoped
-public class ManagementAttachmentCache extends AbstractAttachmentResourceContainer {
+public class ManagementAttachmentResourceContainer extends AbstractAttachmentResourceContainer {
 
   private static final long serialVersionUID = -4688810253543996810L;
