+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/management/resources/ManagementShortlinkResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.management.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.management.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractShortlinkResourceContainer;
 
 @ApplicationScoped
-public class ManagementShortlinkCache extends AbstractShortlinkResourceContainer {
+public class ManagementShortlinkResourceContainer extends AbstractShortlinkResourceContainer {
 
   private static final long serialVersionUID = -5315510394261705312L;
